+++ src/widgets/world/RenderWindow.cpp
@@ -139,6 +139,16 @@ struct RenderWindow::Private {
         if (settingsWidget->pathTracingSchedule() == "Wavefront") {
           options.setTracingBackend(settingsWidget->wavefrontTracingBackend().toStdString());
         }
+        options.setTracingExecution(settingsWidget->tracingExecution().toLower().toStdString());
+        if (settingsWidget->tracingExecution() == "CPU") {
+          options.setIntersectionBackend("cpu");
+        } else if (settingsWidget->tracingExecution() == "GPU") {
+          options.setIntersectionBackend("gpu");
+        } else if (settingsWidget->pathTracingSchedule() == "Wavefront" &&
+                   settingsWidget->tracingExecution() == "Hybrid") {
+          options.setIntersectionBackend(
+            settingsWidget->wavefrontIntersectionBackend().toStdString());
+        }
       }
       options.setMaximumThreads(settingsWidget->renderThreads());
       options.setQueueSize(settingsWidget->queueSize());